extracted dates into it's own file, to make it easier to navigate the pkg
