+++ dinero.go
@@ -12,7 +12,6 @@ import (
 	"net/url"
 	"strconv"
 	"strings"
-	"time"
 )
 
 const (
@@ -55,52 +54,11 @@ type PaginationResult struct {
 	Page                int
 }
 
-// Time is a wrapper for time to make sure the
-// JSON returned from dinero is correctly parsed
-type Time struct {
-	time.Time
-}
-
-const dineroLayout = "2006-01-02T15:04:05.999"
-
-// UnmarshalJSON is Helper function to parse the Timestamp from dinero
-func (dt Time) UnmarshalJSON(b []byte) error {
-	s := strings.Trim(string(b), "\"")
-	if s == "null" {
-		dt.Time = time.Time{}
-		return nil
-	}
-
-	t, err := time.Parse(dineroLayout, s)
-	if err != nil {
-		return err
-	}
-
-	dt.Time = t
-
-	return nil
-}
-
-// MarshalJSON is Helper function to format a dinero timestamp
-func (dt Time) MarshalJSON() ([]byte, error) {
-	return []byte(dt.Time.Format(dineroLayout)), nil
-}
-
-// Date is a representation of dinero´s wanted date format, YYYY-MM-DD
-type Date string
-
-const dineroDateFormat = "2006-01-02"
-
-// NewDate returns the date in correct format for dinero to accept
-func NewDate(year, month, day int) Date {
-	t := time.Date(year, time.Month(month), day, 0, 0, 0, 0, time.UTC)
-	return Date(t.Format(dineroDateFormat))
-}
-
-// DateNow returns the current date in UTC
-func DateNow() Date {
-	t := time.Now().UTC()
-	return Date(t.Format(dineroDateFormat))
+// TimestampResponse is a generic response that is used to return the ID and timestamp from dinero
+// whenever a succesful action has occured
+type TimestampResponse struct {
+	ID        string `json:"guid,omitempty"`
+	Timestamp string `json:"timestamp,omitempty"`
 }
 
 // Authorize will authorize the client
